requirements.txt
kafka-python>=2.0.2
pyod>=1.1.0
pandas>=1.5.0
numpy>=1.23.0
psycopg2-binary>=2.9.5
streamlit>=1.18.0
streamlit-option-menu>=0.3.2
streamlit-autorefresh>=0.0.2
scikit-learn>=1.2.0
matplotlib>=3.6.0
altair>=4.2.0
pickle-mixin>=1.0.2
